rename image_processing to data_processing
