+++ scripts/connect-lines
@@ -10,7 +10,7 @@ from sea_ice_rs.utils import (
     process_single_input,
     process_multiple_inputs,
 )
-from sea_ice_rs.image_processing import connect_lines
+from sea_ice_rs.data_processing import connect_lines
 
 
 def main(args):
