+++ scripts/threshold
@@ -6,7 +6,7 @@ install()
 import argparse
 import sys
 import os
-from sea_ice_rs.image_processing import threshold
+from sea_ice_rs.data_processing import threshold
 from sea_ice_rs.utils import (
     process_single_input,
     process_multiple_inputs,
